requirements.txt
discord.py
frozenmap